feat: add styles for header
