+++ react-performance/src/pages/MainPage.tsx
@@ -3,7 +3,6 @@ import CardList from '../components/widgets/CardList';
 import { useSearchQuery } from '../hooks/useSearchQuery';
 
 const MainPage = () => {
-  // const [searchValue, setSearchValue] = useState<string>(localStorage.getItem('searchTerm') || '');
   const { searchTerm, setSearchTerm } = useSearchQuery();
 
   const handleSearchUpdate = (newSearchTerm: string) => {
@@ -13,8 +12,10 @@ const MainPage = () => {
 
   return (
     <>
-      <Header onSearchChange={handleSearchUpdate} />
-      <CardList />
+      <div className="main-container">
+        <Header onSearchChange={handleSearchUpdate} />
+        <CardList />
+      </div>
     </>
   );
 };
